Moving admin warning tests
